% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{amalgamated_fits_corHMM}
\alias{amalgamated_fits_corHMM}
\title{A wrapper for corHMM that fits Structured Hidden State Models.}
\usage{
amalgamated_fits_corHMM(td, amalgamations, ...)
}
\arguments{
\item{td}{A treeplyr treedata object}

\item{amalgamations}{A list produced by `amalgamate_deps`}

\item{...}{Additional arguments passed to `corHMM::corHMM`}
}
\value{
A list of fits for each character in the dataset
}
\description{
A wrapper for corHMM that fits Structured Hidden State Models.
}
\details{
This function fits the models contained in `amalgamations$M`
to each of the characters in `td` using the function `corHMM::corHMM`.
}
