% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{amalgamated_simmaps_phytools}
\alias{amalgamated_simmaps_phytools}
\title{A wrapper for phytools that draws stochastic character maps given a corHMM fit.}
\usage{
amalgamated_simmaps_phytools(fits, ...)
}
\arguments{
\item{fits}{A list produced by `amalgamated_fits_corHMM`}

\item{...}{Additional arguments passed to `phytools::make.simmap`}
}
\value{
A list of stochastic character maps for each character in the dataset
}
\description{
A wrapper for phytools that draws stochastic character maps given a corHMM fit.
}
\details{
This function takes the models produced from `amalgamated_fits_corHMM` and
uses them to generate stochastic character maps.
}
