% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{amaSMM}
\alias{amaSMM}
\alias{amaSMM_2Q}
\title{SMM amalgamation of rate matrices.}
\usage{
amaSMM(
  ...,
  controlling.state = NULL,
  name.sep = "",
  diag.as = NULL,
  non.rate.as = NULL
)

amaSMM_2Q(
  Qc,
  Qd,
  controlling.state = NULL,
  name.sep = "",
  diag.as = NULL,
  non.rate.as = NULL
)
}
\arguments{
\item{...}{rate matrices.}

\item{controlling.state}{state(s) of Qc that switches on/off Qd matrix.}

\item{name.sep}{separator for the state names.}

\item{diag.as}{populates main diagonal elements.}

\item{non.rate.as}{changes negative elements in the returned matrix to the specified value.}

\item{Qc}{rate matrix (controlling character if dependency is true).}

\item{Qd}{rate matrix (dependent character); if dependency is true then Qd depends on those states of Qc specified in controlling.state.}
}
\value{
matrix
}
\description{
Amalgamating two (\code{amaSMM_2Q}) or several (\code{amaSMM}) rate matrices as independently (SMM-ind) or dependently (SMM-sw) evolving (Tarasov 2019, 2020). The parametric scheme of matrices is defined by integers, different integers
indicate different rate parameters.
}
\examples{
C1 <-matrix(c(-1, 1, 1, -1), 2, 2, byrow = TRUE, dimnames =list( c("r", "b"), c("r", "b")) )
C2 <-matrix(c(-1, 1, 1, -1), 2, 2, byrow = TRUE, dimnames =list( c("o", "g"), c("o", "g")) )
C3 <-matrix(c(-1, 1, 1, -1), 2, 2, byrow = TRUE, dimnames =list( c("w", "m"), c("w", "m")) )
amaSMM(C1, C2, C3)
Tl <- matrix(c(-1, 1, 1, -1), 2, 2, byrow = TRUE, dimnames =list( c("a", "p"), c("a", "p")) )
C <-matrix(c(-2, 2, 2, -2), 2, 2, byrow = TRUE, dimnames =list( c("r", "b"), c("r", "b")) )
amaSMM_2Q(Tl, C, controlling.state = NULL)
amaSMM_2Q(Tl, C, controlling.state = 2)
amaSMM(Tl, C, controlling.state = NULL)
}
\references{
Tarasov, S., 2019. Integration of anatomy ontologies and evo-devo using structured Markov models suggests a new framework for modeling discrete phenotypic traits. Systematic biology, 68(5), pp.698-716.
(\href{https://academic.oup.com/sysbio/article/68/5/698/5298740?login=true}{Read})

Tarasov, S., 2020. The invariant nature of a morphological character and character state: insights from gene regulatory networks. Systematic biology, 69(2), pp.392-400.
(\href{https://academic.oup.com/sysbio/article/69/2/392/5541792?login=true}{Read})
}
\author{
Sergei Tarasov
}
