% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{as_matrixRB}
\alias{as_matrixRB}
\title{Converts rate matrix Q for using in RevBayes.}
\usage{
as_matrixRB(Q, round = 2, symb = NULL)
}
\arguments{
\item{Q}{rate matrix.}

\item{round}{rate rounding.}

\item{symb}{symbol for rates.}
}
\value{
text string
}
\description{
Produces rate matrix that can be copy/pasted into RevBayes script. The rate matrix can be represented as integers or symbols.
}
\examples{
Tl <- matrix(c(-1, 1, 1, -1), 2, 2, byrow = TRUE, dimnames =list( c("T*", "T"), c("T*", "T")) )
C <-matrix(c(-2, 2, 2, -2), 2, 2, byrow = TRUE, dimnames =list( c("r", "b"), c("r", "b")) )
Q <- amaSMM(Tl, C)
#
rb1 <- as_matrixRB(Q)
cat(rb1)
rb2 <- as_matrixRB(Q, symb='q')
cat(rb2)
}
\author{
Sergei Tarasov
}
