% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{as_matrixRB_numeric}
\alias{as_matrixRB_numeric}
\title{Converts rate matrix Q for using in RevBayes (entities as numbers).}
\usage{
as_matrixRB_numeric(Q, round = 2)
}
\description{
Converts rate matrix Q for using in RevBayes (entities as numbers).
}
\author{
Sergei Tarasov
}
