% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{as_matrixRB_symbolic}
\alias{as_matrixRB_symbolic}
\title{Converts rate matrix Q for using in RevBayes (entities as symbols).}
\usage{
as_matrixRB_symbolic(Q, symb = "q")
}
\description{
Converts rate matrix Q for using in RevBayes (entities as symbols).
}
\author{
Sergei Tarasov
}
