% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{auto_dep_matrix}
\alias{auto_dep_matrix}
\title{Obtains a 'auto-dependency' matrix when amalgamating characters referring to the same anatomical entity.}
\usage{
auto_dep_matrix(td, tax.col = FALSE)
}
\arguments{
\item{td}{A treeplyr treedata object with characters to recode}

\item{tax.col}{Indicates if the first column of the data set contains taxon names}
}
\description{
Obtains a 'auto-dependency' matrix when amalgamating characters referring to the same anatomical entity.
}
\examples{
\dontrun{
    auto_dep_matrix(treedata, tax.col = FALSE)
 }

}
