% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{build.chars}
\alias{build.chars}
\title{Build characters based on inferred phenotype clusters.}
\usage{
build.chars(CH, chars.obj)
}
\arguments{
\item{CH}{A named list produced by the function `extract.chars`}

\item{chars.obj}{A data.frame or list produced by the function `chars` from the package 'rphenoscape'}
}
\value{
A list with two main elements: $solved and $unsolved
$solved contains all phenotypes assigned to inferred clusters
$unsolved contains phenotypes that were not assigned to clusters or singletons. Each main element contains:
$chars Inferred clusters of phenotypes with character statements writen in natural language, as in the original study
$clusters Inferred clusters shown as phenotype IDs
$tokens Codes assigned to character states

\dontrun{
    build.chars(CH, chars.obj)
}
}
\description{
Build characters based on inferred phenotype clusters.
}
