% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{build.matrix}
\alias{build.matrix}
\title{Build character matrices based on inferred characters.}
\usage{
build.matrix(tax, pheno.obj, chars.obj, CHARS)
}
\arguments{
\item{chars.obj}{data.frame or list produced by the function 'chars' from the package rphenoscape.}

\item{CHARS}{a named list of inferred characters produced by the function 'build.chars'.}

\item{character, }{a character vector with the taxon names to be scored.}

\item{a}{list of phenotypes produced by the function 'get_phenotypes' and converted with 'as.phenotype' from the package rphenoscape.}
}
\value{
A data.frame with the character matrix.
}
\description{
Build character matrices based on inferred characters.
}
\author{
Diego S. Porto
}
