% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{CHEtransform}
\alias{CHEtransform}
\title{Correlated Hyperspace Extension of rate matrix}
\usage{
CHEtransform(Q, threshold = 10)
}
\arguments{
\item{Q}{rate matrix}

\item{threshold}{rounds the values in its first argument to the specified number of decimal places. It is needed for rate comparing rates using round().}
}
\value{
a list that contains matrix and state names generated by latent characters
}
\description{
Gives minimal Correlated Hyperspace Extension by first using EHEtransform() and then finding hypercube embedding.
So far, it works only for two-state rate matrices. Since there are many minimal CHE transformations for a given Q, it randomly generates one.
}
\examples{
Q <- initQ(c(1, 2), c(.3,.2))
tr <-CHEtransform(Q)

Q.che <-tr$Q
part_scheme=list(c(1:4), c(5:8))
is_slumpable(Q.che, part_scheme)
}
