% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{dep_matrix}
\alias{dep_matrix}
\title{Obtains a dependency matrix when using external ontologies.}
\usage{
dep_matrix(td, ONT, tax.col = F)
}
\arguments{
\item{td}{a treeplyr treedata object with anatomical entities to assess dependencies.}

\item{tax.col}{logical, indicates if the first column of the data set contains taxon names.}

\item{ONTO}{an ontology_index object of an external ontology imported using the R package 'ontologyIndex'.}
}
\value{
A matrix M\[i,j\] where 1 indicates the dependency of the anatomical entity i on j.
}
\description{
Obtains a dependency matrix when using external ontologies.
}
\author{
Diego S. Porto
}
