% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMO_functions.R
\name{discr_Simmap}
\alias{discr_Simmap}
\title{Reading unsummarized Stochastic Map files from RevBayes for one tree.}
\usage{
discr_Simmap(tree, res)
}
\description{
Reading unsummarized Stochastic Map files from RevBayes for one tree.
}
