% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{discr_Simmap_all}
\alias{discr_Simmap_all}
\title{Reading unsummarized Stochastic Map files from RevBayes.}
\usage{
discr_Simmap_all(tree, res)
}
\description{
Reading unsummarized Stochastic Map files from RevBayes.
}
