% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{dropDependentTraits}
\alias{dropDependentTraits}
\title{Utility function for cleaning up character data table after amalgamating characters.}
\usage{
dropDependentTraits(char_info, dep.mat, td)
}
\arguments{
\item{dep.mat}{data.frame, a dependency matrix produced by the function 'pa_dep_matrix' from the package rphenoscape.}

\item{td}{a treeplyr treedata object.}
}
\description{
Utility function for cleaning up character data table after amalgamating characters.
}
