% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{EHEtransform}
\alias{EHEtransform}
\title{Equal Rate Hidden Extension of rate matrix.}
\usage{
EHEtransform(Q)
}
\arguments{
\item{Q}{rate matrix.}
}
\value{
matrix.
}
\description{
Gives minimal equal rate hidden extension by finding least common multiple of rates.
}
\examples{
Q <- initQ(c(1, 2), c(.3,.2))
Qehe <- EHEtransform(Q)

part_scheme=list(c(1, 2), c(3,4,5))
is_slumpable(Qehe, part_scheme)
}
\author{
Sergei Tarasov
}
