% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract.chars}
\alias{extract.chars}
\title{Infer clusters of phenotypes based on mutually exclusivity.}
\usage{
extract.chars(exclu.obj)
}
\arguments{
\item{exclu.obj}{A list produced by the function `mutually_exclusive` from the package 'rphenoscape'}
}
\value{
A list containing:
$phenotypes The inferred clusters of phenotypes (phenotype IDs)
$submatrices The exclusivity submatrices corresponding to each phenotype cluster extracted from the $matrix element of the exclu.obj object
}
\description{
Infer clusters of phenotypes based on mutually exclusivity.
}
\examples{
\dontrun{
    extract.chars(exclu.obj)
}

}
