% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{filter_coverage}
\alias{filter_coverage}
\title{Utility function for filtering data based on missing traits and taxa.}
\usage{
filter_coverage(td, traits = 0, taxa = 0)
}
\arguments{
\item{td}{a treeplyr treedata object.}

\item{traits}{numeric, a threshold value for the coverage across traits.}

\item{taxa}{numeric, a threshold value for the coverage across taxa.}
}
\value{
A treeplyr treedata object filtered by trait and/or taxon coverage.
}
\description{
Utility function for filtering data based on missing traits and taxa.
}
