% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{get_onto_id}
\alias{get_onto_id}
\title{Get IDs for ontology names}
\usage{
get_onto_id(vec_name, ontology, names = F)
}
\arguments{
\item{vec_name}{names od terms}

\item{ontology}{ontology}

\item{names}{use element name}
}
\value{
vector of IDs.
}
\description{
Returns IDs of ontology terms given terms' names
}
\details{
Function to extract IDS from ontology terms.
}
\examples{
vec_name=c("ventral mesofurco-profurcal muscle", "anatomical entity")
get_onto_id(vec_name, HAO)
}
