% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{is_slumpable}
\alias{is_slumpable}
\title{Tests if a rate matrix is strongly lumpable}
\usage{
is_slumpable(Q, part_scheme)
}
\arguments{
\item{Q}{a rate matrix with cells having specific rate values}

\item{part_scheme}{partition of state in Q represented as a list}
}
\value{
Boolean value
}
\description{
Given rate matrix Q and state partition, checks if Q is strongly lumpable
}
\examples{
Q <- initQ(c(1, 2), c(.3,.2))
Q.h <- EHEtransform(Q)

part_scheme=list(c(1, 2), c(3,4,5))
is_slumpable(Q.h, part_scheme)

part_scheme=list(1, c(2, 3, 4, 5))
is_slumpable(Q.h, part_scheme)

Q.h <-CHEtransform(Q)$Q

part_scheme=list(c(1:4), c(5:8))
is_slumpable(Q.h, part_scheme)

part_scheme=list(c(1:3), c(4:8))
is_slumpable(Q.h, part_scheme)
}
