% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{make_tree_eq}
\alias{make_tree_eq}
\title{Make branch lengths and maps equal, use tree.tmp with rounded (3) edge.length.}
\usage{
make_tree_eq(tree.tmp, target.tr, round = 3)
}
\description{
Make branch lengths and maps equal, use tree.tmp with rounded (3) edge.length.
}
