% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{merge.pheno}
\alias{merge.pheno}
\title{Merge compatible phenotypes.}
\usage{
\method{merge}{pheno}(x)
}
\arguments{
\item{x}{A matrix.

Internal function. Not exported.}
}
\description{
Merge compatible phenotypes.
}
