% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{ontologyHeatMap}
\alias{ontologyHeatMap}
\title{Plots a heatmap along with a phylogeny and trait tree.}
\usage{
ontologyHeatMap(td, njt, start = 3, margs = c(0.2, 0.25), ...)
}
\arguments{
\item{td}{The treedata object}

\item{njt}{The trait tree created with the makeTraitTree() function}

\item{start}{Starting column for traits to visualize in the treedata data.frame}

\item{margs}{Margins for the plot}

\item{...}{Extra arguments to pass for visualization}
}
\value{
Plots a heatmap that matches a phylogeny to a trait tree.
}
\description{
Plots a heatmap along with a phylogeny and trait tree.
}
