% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{paramo.list}
\alias{paramo.list}
\title{Final stack of maps for a set of stochastic maps stored in a list.}
\usage{
paramo.list(cc, tree.list, ntrees = 1)
}
\arguments{
\item{cc}{chars id to stack}

\item{tree.list}{Named list with stochastic character maps}

\item{ntrees}{number of trees to stack}
}
\value{
A list of stacked stochastic character maps
}
\description{
Final stack of maps for a set of stochastic maps stored in a list.
}
