% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{prepareMapsRayDISC}
\alias{prepareMapsRayDISC}
\title{Function for processing RevBayes stochastic maps (V2).}
\usage{
prepareMapsRayDISC(td, simmaps, discretization_level = 100)
}
\description{
Function for processing RevBayes stochastic maps (V2).
}
