% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{print_coverage}
\alias{print_coverage}
\title{Utility function for displaying the number of matching species and average value across them from an Ontotrace matrix.}
\usage{
print_coverage(x)
}
\arguments{
\item{x}{data.frame, a data.frame produced by the function 'pk_get_ontotrace' from the package rphenoscape.}
}
\value{
A data.frame with columns displaying entity names, number of matching species, and average value of coverage (percentage).
}
\description{
Utility function for displaying the number of matching species and average value across them from an Ontotrace matrix.
}
