% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{process.ontofast}
\alias{process.ontofast}
\title{Utility function for process an ontoFAST object}
\usage{
process.ontofast(ontoFAST, ONT, s.filter = F, g.filter = F, s.terms, g.terms)
}
\arguments{
\item{ontoFAST}{A named list produced by the function `annot_all_chars` from the package ontoFAST}

\item{s.filter}{Indicates if to use a strict filter of terms}

\item{g.filter}{Indicates if to use a general filter of terms}

\item{s.terms}{A character vector with particular ontology terms to be sorted out}

\item{g.terms}{A character vector with general terms or expressions to be sorted out}

\item{ONTO}{An 'ontology_index' object of an external ontology imported in R using the package 'ontologyIndex'}
}
\description{
Utility function for process an ontoFAST object
}
