% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{read_Simmap_Rev}
\alias{read_Simmap_Rev}
\title{Reading unsummarized Stochastic Map files from RevBayes.}
\usage{
read_Simmap_Rev(file, start = 1, end = 1, save = NULL)
}
\arguments{
\item{file}{file}

\item{start}{start from tree}

\item{end}{end with tree}

\item{save}{save to file. if NULL reads in R}
}
\description{
Reading unsummarized Stochastic Map files from RevBayes.
}
