% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{recode_traits}
\alias{recode_traits}
\title{Amalgamates and recodes traits using the matrices returned by amalgamate_deps.}
\usage{
recode_traits(td, M)
}
\arguments{
\item{td}{a treeplyr treedata object with characters to recode.}

\item{M}{a list produced by the function 'amalgamate_deps'.}
}
\value{
A treeplyr treedata object with characters amalgamated and recoded.
}
\description{
Amalgamates and recodes traits using the matrices returned by amalgamate_deps.
}
