% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{recode_traits_gen}
\alias{recode_traits_gen}
\title{Amalgamates and recodes traits using the matrices returned by amalgamate_deps_gen}
\usage{
recode_traits_gen(td, amal.deps, tax.col = T)
}
\arguments{
\item{td}{A treeplyr treedata object with characters to recode}

\item{amal.deps}{A list produced by the function `amalgamate_deps_gen`}

\item{tax.col}{Indicates if the first column of the data set contains taxon names}
}
\description{
Amalgamates and recodes traits using the matrices returned by amalgamate_deps_gen
}
