% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rescore}
\alias{rescore}
\title{Rescore individual states.}
\usage{
rescore(x, y)
}
\arguments{
\item{x}{A list of taxon phenotypes.}

\item{y}{A list of character state tokens.

Internal function. Not exported.}
}
\description{
Rescore individual states.
}
