% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{sample.ss}
\alias{sample.ss}
\title{Sample pairwise semantic similarity among phenotypes.}
\usage{
sample.ss(exclu.obj, n)
}
\arguments{
\item{exclu.obj}{a list produced by the function 'mutually_exclusive'.}

\item{n}{integer, number of samples.}
}
\value{
A numeric vector of sampled values of pairwise semantic similarity among phenotypes.
}
\description{
Sample pairwise semantic similarity among phenotypes.
}
\author{
Diego S. Porto
}
