% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{strip_IRI}
\alias{strip_IRI}
\title{Utility function for stripping off url from IRI.}
\usage{
strip_IRI(x)
}
\arguments{
\item{x}{character, a character string with an IRI.}
}
\description{
Utility function for stripping off url from IRI.
}
