% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{table2list}
\alias{table2list}
\title{Convert table to list}
\usage{
table2list(table, id_col = c(1), descendants_cols = c(2:ncol(table)))
}
\arguments{
\item{table}{A character table with annotations.}

\item{id_col}{A column ID corresponding to character}

\item{descendants_cols}{IDs of columns corresponding to character annotations}
}
\value{
The list.
}
\description{
Takes table where each row consists of character number and ontology annotations and returns a list.
Each character is assigned its own ID CHAR:XXXX
}
\details{
Function to convert a table of character annotations to list.
}
\examples{
# converting Sharkey_2011 data set to list of characater states
list_data<-table2list(Sharkey_2011)
}
