% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{dep_matrix}
\alias{dep_matrix}
\title{Obtains a presence-absence dependency matrix when using external ontologies}
\usage{
dep_matrix(td, ONT, tax.col = F)
}
\arguments{
\item{td}{A treeplyr treedata object with characters to recode}

\item{tax.col}{Indicates if the first column of the data set contains taxon names}

\item{ONTO}{An 'ontology_index' object of an external ontology imported in R using the package 'ontologyIndex'}
}
\description{
Obtains a presence-absence dependency matrix when using external ontologies
}
