% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dropDependentTraits}
\alias{dropDependentTraits}
\title{Utility function for cleaning up character data table after amalgamating characters.}
\usage{
dropDependentTraits(char_info, dep.mat, td)
}
\arguments{
\item{char_info}{The character data table to clean}

\item{dep.mat}{Dependency matrix}

\item{td}{The treedata object with amalgamated characters}
}
\value{
char_info_comb Cleaned character data table after amalgamating characters.
}
\description{
Utility function for cleaning up character data table after amalgamating characters.
}
