% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{init_binary_matrix}
\alias{init_binary_matrix}
\title{Initialize binary matrices given a graph.}
\usage{
init_binary_matrix(graph)
}
\arguments{
\item{graph}{igraph object.}
}
\value{
List of matrices.
}
\description{
Call matrices are populated with different parameters.
}
\examples{
init_binary_matrix(g)

}
\author{
Sergei Tarasov
}
