% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Discr_maps.R
\name{make_tree_eq}
\alias{make_tree_eq}
\title{make branch length and maps equal, use tree.tmp with rounded (3) edge.length}
\usage{
make_tree_eq(tree.tmp, target.tr, round = 3)
}
\description{
make branch length and maps equal, use tree.tmp with rounded (3) edge.length
}
