% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{makeTraitTree}
\alias{makeTraitTree}
\title{Make a Trait Tree}
\usage{
makeTraitTree(td, external = FALSE, ONT, method = "nj")
}
\arguments{
\item{td}{The treedata object containing character data}

\item{external}{Should an external ontology be used}

\item{ONT}{The ontology object to use}

\item{method}{Neighbor-joining method for clustering-dendrogram. Options are "nj" or "cls"}
}
\description{
Makes a trait tree using semantic similarity.
}
\details{
Makes a trait tree using semantic similarity.
}
\examples{
\dontrun{
    makeTraitTree(treedata, external = TRUE, HAO)
}

}
