% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{makeTraitTree}
\alias{makeTraitTree}
\title{Makes a trait tree using semantic similarity.}
\usage{
makeTraitTree(td, external = F, ONT, method = "nj")
}
\description{
Makes a trait tree using semantic similarity.
}
