% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{paramo}
\alias{paramo}
\title{Final stack of maps for a set of stochastic maps stored in a directory.}
\usage{
paramo(cc, ntrees = 10, dirW = c(""))
}
\arguments{
\item{cc}{chars id to stack}

\item{ntrees}{number of trees to stack}

\item{dirW}{directory for zip file}
}
\value{
A list of stacked stochastic character maps
}
\description{
Final stack of maps for a set of stochastic maps stored in a directory.
}
