% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{prepareMapsRev}
\alias{prepareMapsRev}
\title{Function for processing RevBayes stochastic maps (V1).}
\usage{
prepareMapsRev(td, discretization_level = 100, start_tree = 1, end_tree = 2)
}
\description{
Function for processing RevBayes stochastic maps (V1).
}
