% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print_coverage}
\alias{print_coverage}
\title{Utility function for displaying the number of matching species and average value across them from an Ontotrace matrix.}
\usage{
print_coverage(x)
}
\description{
Utility function for displaying the number of matching species and average value across them from an Ontotrace matrix.
}
