% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{process.ontofast}
\alias{process.ontofast}
\title{Utility function for processing an ontoFAST object.}
\usage{
process.ontofast(ontoFAST, ONT, s.filter = F, g.filter = F, s.terms, g.terms)
}
\arguments{
\item{ontoFAST}{a named list produced by the function 'annot_all_chars' from the package ontoFAST.}

\item{s.filter}{logical, indicates if to use a strict filter of terms.}

\item{g.filter}{logical, indicates if to use a general filter of terms.}

\item{s.terms}{character, a character vector with particular ontology terms to be sorted out.}

\item{g.terms}{character, a character vector with general terms or expressions to be sorted out.}

\item{ONTO}{an ontology_index object of an external ontology imported using the R package 'ontologyIndex'.}
}
\value{
A list with filtered candidate terms from the output of the function 'annot_all_chars' from the package ontoFAST.
}
\description{
Utility function for processing an ontoFAST object.
}
\author{
Diego S. Porto
}
