% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{recode_td}
\alias{recode_td}
\title{Recodes a treedata object based on amalgamated characters.}
\usage{
recode_td(td, traits, states, depstates = numeric(0))
}
\description{
Recodes a treedata object based on amalgamated characters.
}
