% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{recode_traits}
\alias{recode_traits}
\title{Amalgamates traits using the matrices returned by amalgamate_deps.}
\usage{
recode_traits(td, M)
}
\arguments{
\item{td}{A treeplyr treedata object with characters to recode}

\item{M}{A list produced by the function `amalgamate_deps`}
}
\description{
Amalgamates traits using the matrices returned by amalgamate_deps.
}
