% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scate_functions.R
\name{remove_indirect}
\alias{remove_indirect}
\title{Remove indirect links in a dependency graph.}
\usage{
remove_indirect(dependency_matrix)
}
\arguments{
\item{dependency_matrix}{A dependency matrix}
}
\description{
Function for removing redundant dependency edges from our graph.
}
\examples{
\dontrun{
    remove_indirect(dep_mat)
}

}
