% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{strip_IRI}
\alias{strip_IRI}
\title{Utility function for stripping off url from IRI.}
\usage{
strip_IRI(x)
}
\arguments{
\item{x}{A full IRI}
}
\value{
x The IRI with its url stripped.
}
\description{
Utility function for stripping off url from IRI.
}
